Move Gamepieces class to another file
