+++ lib/Cycle.js
@@ -1,4 +1,4 @@
-const Gamepieces = require('./Game.js');
+const Gamepieces = require('./Gamepieces.js');
 
 class Cycle extends Gamepieces {
   constructor(x, y, w, h, direction, color) {
