+++ lib/Disk.js
@@ -1,4 +1,4 @@
-const Gamepieces = require('./Game.js');
+const Gamepieces = require('./Gamepieces.js');
 
 class Disk extends Gamepieces {
   constructor(x, y, w = 10, h = 10, direction, color) {
