+++ lib/Game.js
@@ -2,17 +2,6 @@ const Cycle = require('./Cycle.js');
 const Trail = require('./Trail.js');
 const Disk = require('./Disk.js');
 
-class Gamepieces {
-  constructor (x, y, w, h, direction, color) {
-    this.x = x;
-    this.y = y;
-    this.w = w;
-    this.h = h;
-    this.direction = direction;
-    this.color = color;
-  }
-}
-
 class Game {
   constructor() {
     this.cycles = [];
@@ -185,4 +174,4 @@ class Game {
   }
 }
 
-module.exports = Game, Gamepieces;
+module.exports = Game;
