+++ lib/index.js
@@ -297,6 +297,10 @@ function end(color) {
     gameText[1].innerText = ['In Round 3, you must avoid the discs flying ',
       'throughout the arena. \n \n \n \n'].join('');
     gameText[2].innerText = 'Press the Enter Key to Enter the Grid';
+    gameText[3].setTimeout(function() {
+      gameText[3].addClass(fade).innerText = 'LOADING... Please wait';
+      $('*').off('keyup keydown keypress');
+    }, 3000);
   } else {
     var p1unique = p1Scores.filter(onlyUnique);
     var p2unique = p2Scores.filter(onlyUnique);
